feat: resize and arrange after uploading multiple pictures to the whiteboard (#8529)
Close issue [BS-1539](https://linear.app/affine-design/issue/BS-1539).

### What Changed?
- Refactor `EdgelessAlignButton` method to `autoArrangeElements` and `autoResizeElements` commands
- Refactor `EdgelessAlignButton` method to `updateXYWH` utils
- Resize and arrange after uploading multiple pictures

<div class='graphite__hidden'>
          <div>🎥 Video uploaded on Graphite:</div>
            <a href="https://app.graphite.dev/media/video/sJGviKxfE3Ap685cl5bj/cbebfc6f-39ba-4272-b115-954aa44413e7.mov">
              <img src="https://app.graphite.dev/api/v1/graphite/video/thumbnail/sJGviKxfE3Ap685cl5bj/cbebfc6f-39ba-4272-b115-954aa44413e7.mov">
            </a>
          </div>
<video src="https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/sJGviKxfE3Ap685cl5bj/cbebfc6f-39ba-4272-b115-954aa44413e7.mov">录屏2024-10-15 17.58.53.mov</video>
