+++ packages/affine/block-surface/src/commands/index.ts
@@ -1,7 +1,13 @@
 import type { BlockCommands } from '@blocksuite/block-std';
 
+import {
+  autoArrangeElementsCommand,
+  autoResizeElementsCommand,
+} from './auto-align.js';
 import { reassociateConnectorsCommand } from './reassociate-connectors.js';
 
 export const commands: BlockCommands = {
   reassociateConnectors: reassociateConnectorsCommand,
+  autoArrangeElements: autoArrangeElementsCommand,
+  autoResizeElements: autoResizeElementsCommand,
 };
